pull in dev

+++ src/Orchard.Web/Modules/Orchard.Taxonomies/Controllers/TermAdminController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Web.Mvc;
 using Orchard.Taxonomies.Models;
 using Orchard.Localization;
+using Orchard.Mvc.Html;
 using Orchard.Taxonomies.ViewModels;
 using Orchard.Taxonomies.Services;
 using Orchard.UI.Admin;
@@ -41,7 +42,7 @@ namespace Orchard.Taxonomies.Controllers {
 
             var taxonomy = _taxonomyService.GetTaxonomy(taxonomyId);
 
-            var terms = _taxonomyService.GetTermsQuery(taxonomyId).Slice(pager.GetStartIndex(), pager.PageSize);
+            var terms = TermPart.Sort(_taxonomyService.GetTermsQuery(taxonomyId).Slice(pager.GetStartIndex(), pager.PageSize));
 
             var pagerShape = Shape.Pager(pager).TotalItemCount(_taxonomyService.GetTermsQuery(taxonomyId).Count());
 
@@ -70,15 +71,24 @@ namespace Orchard.Taxonomies.Controllers {
             var checkedEntries = viewModel.Terms.Where(t => t.IsChecked).ToList();
             switch (viewModel.BulkAction) {
                 case TermsAdminIndexBulkAction.None:
+                    Services.Notifier.Information(T("No action selected."));
                     break;
                 case TermsAdminIndexBulkAction.Delete:
                     if (!Services.Authorizer.Authorize(Permissions.ManageTerms, T("Couldn't delete term")))
                         return new HttpUnauthorizedResult();
+                    
+                    if(!checkedEntries.Any()) {
+                        Services.Notifier.Information(T("No terms selected."));
+                        break;
+                    }
 
                     foreach (var entry in checkedEntries) {
                         var term = _taxonomyService.GetTerm(entry.Id);
                         _taxonomyService.DeleteTerm(term);
                     }
+
+                    Services.Notifier.Information(T.Plural("{0} term has been removed.", "{0} terms have been removed.", checkedEntries.Count));
+
                     break;
                 case TermsAdminIndexBulkAction.Merge:
                     if (!Services.Authorizer.Authorize(Permissions.ManageTerms, T("Couldn't delete term")))
@@ -99,8 +109,6 @@ namespace Orchard.Taxonomies.Controllers {
                 default:
                     throw new ArgumentOutOfRangeException();
             }
-
-            Services.Notifier.Information(T("{0} term have been removed.", checkedEntries.Count));
 
             return RedirectToAction("Index", new { taxonomyId = viewModel.TaxonomyId });
         }
@@ -200,14 +208,14 @@ namespace Orchard.Taxonomies.Controllers {
             }
 
             Services.ContentManager.Publish(term.ContentItem);
-            Services.Notifier.Information(T("The {0} term has been created.", term.Name));
+            Services.Notifier.Success(T("The {0} term has been created.", term.Name));
 
             return RedirectToAction("Index", "TermAdmin", new { taxonomyId });
         }
 
         public ActionResult Edit(int id) {
 
-            if (!Services.Authorizer.Authorize(Permissions.ManageTerms, T("Not allowed to manage taxonomies")))
+            if (!Services.Authorizer.Authorize(Permissions.ManageTerms, T("Not allowed to manage terms")))
                 return new HttpUnauthorizedResult();
 
             var term = _taxonomyService.GetTerm(id);
@@ -220,7 +228,7 @@ namespace Orchard.Taxonomies.Controllers {
 
         [HttpPost, ActionName("Edit")]
         public ActionResult EditPost(int id) {
-            if (!Services.Authorizer.Authorize(Permissions.ManageTaxonomies, T("Couldn't edit taxonomy")))
+            if (!Services.Authorizer.Authorize(Permissions.ManageTerms, T("Couldn't edit term")))
                 return new HttpUnauthorizedResult();
 
             var term = _taxonomyService.GetTerm(id);
@@ -238,7 +246,7 @@ namespace Orchard.Taxonomies.Controllers {
 
             Services.ContentManager.Publish(contentItem);
             _taxonomyService.ProcessPath(term);
-            Services.Notifier.Information(T("Term information updated"));
+            Services.Notifier.Success(T("Term information updated"));
 
             return RedirectToAction("Index", "TermAdmin", new { taxonomyId = term.TaxonomyId });
         }
